fix: 임시 메모 태그 불러오는 오류 수정

+++ src/main/java/Funssion/Inforum/domain/post/memo/repository/MemoRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 public interface MemoRepository {
 
     Memo create(Memo memo);
-    List<Memo> findAllByDaysOrderByLikes(Long days);
+    List<Memo> findAllByDaysOrderByLikes(Integer days);
     List<Memo> findAllOrderById();
     List<Memo> findAllByUserIdOrderById(Long userId);
     List<Memo> findAllLikedMemosByUserId(Long userId);
